backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy
mediapipe
opencv-python-headless
pymongo
python-dotenv